fix(blog): 修正数据库的部分int到bigint

+++ blog/module-admin/src/main/java/top/fxmarkbrown/blog/controller/article/SysArticleController.java
@@ -32,7 +32,7 @@ public class SysArticleController {
 
     @GetMapping("/detail/{id}")
     @Operation(summary = "文章详情")
-    public Result<SysArticleDetailVo> detail(@PathVariable Integer id) {
+    public Result<SysArticleDetailVo> detail(@PathVariable Long id) {
         return Result.success(sysArticleService.detail(id));
     }
